Use proper name when checking if container exists
There is a nice, human-friendly project name defined in .spark.yml that is passed to Docker Compose, which lowercases it and removes all spaces to create a prefix that is used for naming the containers Docker Compose orchestrates.

We missed removing spaces, which this PR fixes. It also introduces a new method to retrieve the appropriate container name, which we can reuse if needed.