Adjust Composer script name: robo → spark
The new readme includes adding a Composer script with the name ‘spark’. When Spark was originally written this name was ‘robo’, and that’s how our tasks which we use to execute others referred to it. This breaks things if a Sparkfied project is set up according to the new readme, which makes more sense, so let’s adjust how we invoke the Composer script.